fix(folder-structure): seperate pages from components
